requirements.txt
imutils==0.5.3
numpy==1.18.0
opencv-contrib-python==4.4.0.42
Pillow==7.2.0